Update: Move navi-request-column-config -> navi-column-config
