+++ packages/reports/addon/components/navi-column-config/dimension.js
@@ -1,11 +1,11 @@
 /**
- * Copyright 2019, Yahoo Holdings Inc.
+ * Copyright 2020, Yahoo Holdings Inc.
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
  *
- * Description: Navi Request Column Config Base Component
+ * Description: Navi Request Column Config Dimension Component
  *
  * Usage:
- *  <NaviRequestColumnConfig::Base
+ *  <NaviRequestColumnConfig::Dimension
  *    @column={{editingColumn}}
  *    @metadata={{visualization.metadata}}
  *    @onClose={{action "onClose"}}
@@ -14,10 +14,10 @@
  */
 import Component from '@ember/component';
 import { layout as templateLayout, tagName } from '@ember-decorators/component';
-import layout from '../../templates/components/navi-request-column-config/base';
+import layout from '../../templates/components/navi-column-config/dimension';
 
 @tagName('')
 @templateLayout(layout)
-class Base extends Component {}
+class Dimension extends Component {}
 
-export default Base;
+export default Dimension;
