+++ packages/reports/addon/components/navi-column-config/metric.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019, Yahoo Holdings Inc.
+ * Copyright 2020, Yahoo Holdings Inc.
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
  *
  * Description: Navi Request Column Config Metric Component
@@ -14,7 +14,7 @@
  */
 import Component from '@ember/component';
 import { layout as templateLayout, tagName } from '@ember-decorators/component';
-import layout from '../../templates/components/navi-request-column-config/metric';
+import layout from '../../templates/components/navi-column-config/metric';
 import { action, set, computed } from '@ember/object';
 import { inject as service } from '@ember/service';
