+++ packages/reports/addon/templates/components/navi-column-config/base.hbs
