+++ packages/reports/addon/templates/components/navi-column-config/dimension.hbs
