+++ packages/reports/addon/templates/components/navi-column-config/metric.hbs
