+++ packages/reports/app/components/navi-column-config/date-time.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/components/navi-column-config/date-time';
