--- packages/reports/app/components/navi-request-column-config/metric.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/components/navi-request-column-config/metric';
